refactor(benchmarks): extract VLMBias scoring utils
