+++ evalscope/benchmarks/vlms_are_biased/vlms_are_biased_adapter.py
@@ -10,6 +10,8 @@ from evalscope.api.metric.scorer import Score
 from evalscope.api.registry import register_benchmark
 from evalscope.constants import Tags
 
+from .utils import normalize_answer, score_answer
+
 SUBSET_LIST = [
     'main',
     'identification',
@@ -49,39 +51,6 @@ VLMs Are Biased (VLMBias) evaluates whether vision-language models answer object
 - Generation should be deterministic and concise; the official lmms-eval setup uses `temperature=0` and at most 32 new tokens
 - [Paper](https://arxiv.org/abs/2505.23941) | [GitHub](https://github.com/anvo25/vlms-are-biased) | [Project page](https://vlmsarebiased.github.io/)
 """
-
-
-def normalize_answer(answer: str) -> str:
-    """Normalize an answer using the official lmms-eval comparison rule."""
-    return str(answer or '').strip().lower().strip('{}').strip()
-
-
-def score_answer(prediction: str, ground_truth: str, expected_bias: Any = None) -> Dict[str, float]:
-    """Compute the official accuracy and optional bias-ratio indicators."""
-    normalized_prediction = normalize_answer(prediction)
-    normalized_ground_truth = normalize_answer(ground_truth)
-    normalized_bias = normalize_answer(expected_bias) if expected_bias is not None else None
-
-    is_correct = normalized_prediction == normalized_ground_truth
-    matches_bias = normalized_bias is not None and normalized_prediction == normalized_bias
-
-    if not is_correct and not matches_bias:
-        prediction_digits = ''.join(character for character in normalized_prediction if character.isdigit())
-        ground_truth_digits = ''.join(character for character in normalized_ground_truth if character.isdigit())
-        bias_digits = (
-            ''.join(character for character in normalized_bias if character.isdigit())
-            if normalized_bias is not None
-            else ''
-        )
-        if prediction_digits and ground_truth_digits:
-            is_correct = prediction_digits == ground_truth_digits
-        if prediction_digits and bias_digits:
-            matches_bias = prediction_digits == bias_digits
-
-    result = {'acc': float(is_correct)}
-    if normalized_bias is not None:
-        result['bias_ratio'] = float(matches_bias)
-    return result
 
 
 @register_benchmark(
